Fixed some errors
