--- src/main/java/org/example/Building/Shop/Drinks.java
@@ -1,4 +1,0 @@
-package org.example.Building.Shop;
-
-public class Drinks {
-}
