--- src/main/java/org/example/Building/Shop/Food.java
@@ -1,4 +1,0 @@
-package org.example.Building.Shop;
-
-public class Food {
-}
