--- src/main/java/org/example/Building/Shop/Gifts.java
@@ -1,5 +1,0 @@
-package org.example.Building.Shop;
-
-public class Gifts {
-
-}
