--- src/main/java/org/example/Building/Shop/Tickets.java
@@ -1,4 +1,0 @@
-package org.example.Building.Shop;
-
-public class Tickets {
-}
