+++ src/main/java/org/example/Buildings/Enclosure.java
@@ -1,11 +1,15 @@
-package org.example.Building;
+package org.example.Buildings;
+
+import org.example.Animal;
+import org.example.Building;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Scanner;
 
 public abstract class Enclosure extends Building {
     private List<Animal> animal = new ArrayList<>();
-    public Scanner scanner = new scanner(System.in);
+    public Scanner scanner = new Scanner(System.in);
 
     public void openEnclosure(){
