+++ src/main/java/org/example/Buildings/Shops/Drinks.java
@@ -1,0 +1,4 @@
+package org.example.Buildings.Shops;
+
+public class Drinks {
+}
