+++ src/main/java/org/example/Buildings/Shops/Food.java
@@ -1,0 +1,4 @@
+package org.example.Buildings.Shops;
+
+public class Food {
+}
