+++ src/main/java/org/example/Buildings/Shops/Gifts.java
@@ -1,0 +1,5 @@
+package org.example.Buildings.Shops;
+
+public class Gifts {
+
+}
