+++ src/main/java/org/example/Buildings/Shops/Tickets.java
@@ -1,0 +1,6 @@
+package org.example.Buildings.Shops;
+
+import org.example.Buildings.Shop;
+
+public class Tickets {
+}
