+++ src/main/java/org/example/People/Handler.java
@@ -1,5 +1,6 @@
 package org.example.People;
 
+import org.example.Building;
 import org.example.Person;
 
 public class Handler extends Person {
